requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-jwt==1.6.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0